Vision Mamba/train_vision_mamba.py
import torch
import torch.nn as nn
import torch.optim as optim

# A simple wrapper layout for the Vision Mamba (Vim) model setup for Glaucoma Detection
class GlaucomaVim(nn.Module):
    def __init__(self, num_classes=2):
        '''
        Glaucoma classification typically has 2 classes: Normal, Glaucoma.
        '''
        super(GlaucomaVim, self).__init__()
        
        # In a real implementation, you would import the Vim model from its official module:
        # from vim.models_vim import VisionMamba
        # self.vss_blocks = VisionMamba(num_classes=num_classes, ...)
        
        print("Initializing Vision Mamba base architecture for Glaucoma screening...")
        self.patch_embed = nn.Sequential(
            nn.Conv2d(3, 64, kernel_size=3, stride=2, padding=1),
            nn.ReLU()
        )
        # Placeholder for Vision State Space (Mamba) Blocks
        self.vss_blocks = nn.Sequential(
            nn.AdaptiveAvgPool2d((1, 1)),
            nn.Flatten()
        )
        self.head = nn.Linear(64, num_classes)

    def forward(self, x):
        x = self.patch_embed(x)
        features = self.vss_blocks(x)
        return self.head(features)

if __name__ == "__main__":
    # Create the model targeting 2 classes: Normal vs Glaucoma
    model = GlaucomaVim(num_classes=2)
    device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
    model.to(device)

    # Dummy input representing a 224x224 RGB image
    dummy_input = torch.randn(8, 3, 224, 224).to(device) # Batch size of 8
    output = model(dummy_input)
    
    criterion = nn.CrossEntropyLoss()
    optimizer = optim.AdamW(model.parameters(), lr=0.001)

    # Dummy Pass
    optimizer.zero_grad()
    loss = criterion(output, torch.ones(8, dtype=torch.long).to(device))
    loss.backward()
    optimizer.step()

    print("Output shape (Batch Size, Classes):", output.shape)
    print("Vision Mamba setup for Glaucoma initialized and successfully ran a dummy forward-backward pass!")
