requirements.txt
ultralytics
gradio
opencv-python-headless
matplotlib
Pillow
datasets
huggingface_hub
